Se corrigio la estructura para dejar validaciones de rol en paginas en el lado del servidor por medio de cookie (Faltaba a la app de /configuration/pde)
